Added webpack and modified logic to function with it
