+++ src/dom.js
@@ -4,8 +4,7 @@ function render(board) {
     tableRows += '<tr class="table-rows">';
     for (let j = 0; j <= 2; j += 1) {
       if (board.getValue(i * 3 + j) === 0) {
-        tableRows += `<th class="board-border" onclick="space(${i * 3
-      + j})"></th>`;
+        tableRows += `<th class="board-border" id="space-${i * 3 + j}"></th>`;
       } else if (board.getValue(i * 3 + j) === 1) {
         tableRows += '<th class="board-border text-center background-X"></th>';
       } else {
@@ -17,23 +16,23 @@ function render(board) {
   document.getElementById('table-rows').innerHTML = tableRows;
 }
 
-function renderWinnerBorder(board) {
-  let tableRows = '';
-  for (let i = 0; i <= 2; i += 1) {
-    tableRows += '<tr class="table-rows">';
-    for (let j = 0; j <= 2; j += 1) {
-      if (board.getValue(i * 3 + j) === 0) {
-        tableRows += '<th class="board-border"></th>';
-      } else if (board.getValue(i * 3 + j) === 1) {
-        tableRows += '<th class="board-border text-center background-X"></th>';
-      } else {
-        tableRows += '<th class="board-border text-center background-O"></th>';
-      }
-    }
-    tableRows += '</tr>';
-  }
-  document.getElementById('table-rows').innerHTML = tableRows;
-}
+// function renderWinnerBorder(board) {
+//   let tableRows = '';
+//   for (let i = 0; i <= 2; i += 1) {
+//     tableRows += '<tr class="table-rows">';
+//     for (let j = 0; j <= 2; j += 1) {
+//       if (board.getValue(i * 3 + j) === 0) {
+//         tableRows += '<th class="board-border"></th>';
+//       } else if (board.getValue(i * 3 + j) === 1) {
+//         tableRows += '<th class="board-border text-center background-X"></th>';
+//       } else {
+//         tableRows += '<th class="board-border text-center background-O"></th>';
+//       }
+//     }
+//     tableRows += '</tr>';
+//   }
+//   document.getElementById('table-rows').innerHTML = tableRows;
+// }
 
 function openModel() {
   document.getElementById('model').style.display = 'block';
@@ -50,3 +49,7 @@ function closeModel() {
 function closeModelAI() {
   document.getElementById('modelAI').style.display = 'none';
 }
+
+export {
+  render, openModel, openModelAI, closeModel, closeModelAI,
+};
